+++ examples/inflammatory_message/point_intercept.py
@@ -120,7 +120,6 @@ config = Config(
         ),  # type: ignore
         environment=EnvironmentConfig(
             start_tick=6 * 60 * 60,
-            total_tick=18 * 60 * 60,
         ),
     ),
 )